[Improvement] Dependency database memory usage reduced slightly
 - save 24 bytes per node by avoiding wasted bytes due to padding and oversized enum types

+++ Code/Tools/FBuild/FBuildCore/Graph/FileNode.cpp
@@ -16,7 +16,7 @@
 
 // CONSTRUCTOR
 //------------------------------------------------------------------------------
-FileNode::FileNode( const AString & fileName, uint32_t controlFlags )
+FileNode::FileNode( const AString & fileName, uint8_t controlFlags )
     : Node( fileName, Node::FILE_NODE, controlFlags )
 {
     ASSERT( fileName.EndsWith( "\\" ) == false );
